pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.7",
    "langchain-groq>=1.1.1",
    "load-dotenv>=0.1.0",
    "numpy>=2.2.6",
    "output>=1.0.1",
    "scikit-learn==1.6.1",
    "xgboost>=3.1.3",
]
